PHONE: Changed how the languages are fetched in details
